+++ script/del/index.js
@@ -1,0 +1,9 @@
+const del = require('del');
+const fs = require('fs-extra');
+function main() {
+  del('./log');
+
+  fs.copySync('./lib', './dist');
+}
+
+main();