+++ src/webpack-plugin/check-long-constant/index.ts
@@ -1,0 +1,1 @@
+export { CheckLongConstantPlugin } from './check-long-constant';