+++ src/webpack-plugin/check-long-constant/types.ts
@@ -1,0 +1,5 @@
+export interface FoundItem {
+  value: string;
+  name: string;
+  start: number;
+}