+++ src/webpack-plugin/index.ts
@@ -1,5 +1,6 @@
 export * from './add-placeholder';
 export * from './analyze-deps';
+export * from './check-long-constant';
 export * from './copy-dir';
 export * from './custom-preload';
 export * from './dispatch-script';
@@ -15,6 +16,7 @@ export * from './modify-html';
 export * from './rem-to-rpx';
 export * from './replace-app-id';
 export * from './replace-content';
+export * from './replace-uni-h5';
 export * from './save-loader-log';
 export * from './transfer-local-file';