+++ src/webpack-plugin/replace-uni-h5/index.ts
@@ -1,0 +1,1 @@
+export { ReplaceUniH5Plugin } from './replace-uni-h5';