+++ src/webpack-plugin/replace-uni-h5/types.ts
@@ -1,0 +1,5 @@
+export interface IReplaceUniH5PluginOption {
+  // uni-h5 版本，比如 v2.0.1-alpha-36920221111001
+  // 找不到对应的文件时，插件使用失败
+  version: String;
+}